+++ node_launcher/gui/menu/__init__.py
@@ -1,0 +1,2 @@
+from node_launcher.gui.menu.manage_lnd.lnd_output_widget import LndOutputWidget
+from node_launcher.gui.menu.manage_lnd.restart_layout import RestartLayout
