+++ node_launcher/gui/menu/manage_bitcoind/bitcoind_configuration_tab.py
@@ -1,9 +1,9 @@
 from PySide2.QtCore import Qt, Signal
 from PySide2.QtWidgets import QWidget, QLabel, QCheckBox, QVBoxLayout
 
+from node_launcher.constants import Network, MAINNET, TESTNET
+from node_launcher.gui.menu.manage_bitcoind.data_directories import DataDirectoryBox
 
-from node_launcher.constants import Network
-from node_launcher.gui.system_tray_widgets.settings.data_directories.data_directory_box import DataDirectoryBox
 from node_launcher.node_set.bitcoin import Bitcoin
