+++ node_launcher/gui/menu/manage_bitcoind/bitcoind_output_tab.py
