+++ node_launcher/gui/menu/manage_bitcoind/data_directories/data_directory_box.py
