+++ node_launcher/gui/menu/manage_bitcoind/data_directories/datadir_label.py
