+++ node_launcher/gui/menu/manage_bitcoind/data_directories/prune_warning_label.py
