+++ node_launcher/gui/menu/manage_lnd/lnd_manager_tabs_dialog.py
