+++ node_launcher/gui/menu/manage_lnd/restart_layout.py
