--- node_launcher/gui/system_tray_widgets/__init__.py
@@ -1,21 +1,0 @@
-from .advanced import (
-    AdvancedWidget,
-    ConfigurationFilesLayout,
-    PortsLayout,
-    TlsLayout,
-    VersionsLayout,
-    ZapLayout
-)
-from .settings import (
-    DataDirectoryBox,
-    DatadirLabel,
-    PruneWarningLabel,
-    SelectDirectoryDialog,
-    AliasLayout,
-    BitcoinTab,
-    LndTab,
-    SettingsTabDialog
-)
-from node_launcher.gui.system_tray_widgets.manage_bitcoind.bitcoind_output_widget import BitcoindOutputWidget
-from node_launcher.gui.system_tray_widgets.manage_lnd.lnd_output_widget import LndOutputWidget
-from.restart_layout import RestartLayout
