--- node_launcher/gui/system_tray_widgets/advanced/__init__.py
@@ -1,6 +1,0 @@
-from .advanced_widget import AdvancedWidget
-from .configuration_files_layout import ConfigurationFilesLayout
-from .ports_layout import PortsLayout
-from .tls_layout import TlsLayout
-from .versions_layout import VersionsLayout
-from .zap_layout import ZapLayout
