--- node_launcher/gui/system_tray_widgets/settings/__init__.py
@@ -1,9 +1,0 @@
-from .data_directories import (
-    DataDirectoryBox,
-    DatadirLabel,
-    PruneWarningLabel,
-    SelectDirectoryDialog,
-)
-from .alias_layout import AliasLayout
-from .lnd_tab import LndTab
-from .settings_tab_dialog import SettingsTabDialog
