+++ tests/test_gui/test_data_directory/test_data_directory_box.py
@@ -4,7 +4,7 @@ import pytest
 from PySide2.QtCore import Qt
 from PySide2.QtTest import QTest
 
-from node_launcher.gui.system_tray_widgets import DataDirectoryBox
+from node_launcher.gui.menu import DataDirectoryBox
 from node_launcher.node_set import NodeSet
 
 
@@ -15,8 +15,8 @@ def data_directory_box() -> DataDirectoryBox:
     return data_directory_box
 
 
-@patch('node_launcher.gui.system_tray_widgets.settings.data_directories.data_directory_box.SelectDirectoryDialog')
-@patch('node_launcher.gui.system_tray_widgets.settings.data_directories.data_directory_box.reveal')
+@patch('node_launcher.gui.menu.settings.data_directories.data_directory_box.SelectDirectoryDialog')
+@patch('node_launcher.gui.menu.settings.data_directories.data_directory_box.reveal')
 class TestDataDirectoryBox(object):
     def test_show_directory_button(self,
                                    reveal_patch: MagicMock,
