+++ tests/test_gui/test_network_buttons/test_zap_layout.py
@@ -5,7 +5,7 @@ from PySide2.QtCore import Qt
 from PySide2.QtTest import QTest
 
 from node_launcher.constants import TESTNET
-from node_launcher.gui.system_tray_widgets import ZapLayout
+from node_launcher.gui.menu import ZapLayout
 
 
 @pytest.fixture
@@ -17,11 +17,11 @@ def zap_layout() -> ZapLayout:
 
 
 @patch(
-    'node_launcher.gui.system_tray_widgets.advanced.zap_layout.get_deprecated_lndconnect_url')
-@patch('node_launcher.gui.system_tray_widgets.advanced.zap_layout.get_qrcode_img')
-@patch('node_launcher.gui.system_tray_widgets.advanced.zap_layout.webbrowser')
-@patch('node_launcher.gui.system_tray_widgets.advanced.zap_layout.QPixmap')
-@patch('node_launcher.gui.system_tray_widgets.advanced.zap_layout.QLabel')
+    'node_launcher.gui.menu.advanced.zap_layout.get_deprecated_lndconnect_url')
+@patch('node_launcher.gui.menu.advanced.zap_layout.get_qrcode_img')
+@patch('node_launcher.gui.menu.advanced.zap_layout.webbrowser')
+@patch('node_launcher.gui.menu.advanced.zap_layout.QPixmap')
+@patch('node_launcher.gui.menu.advanced.zap_layout.QLabel')
 class TestZapLayout(object):
     def test_open_zap_desktop_button(self,
                                      qlabel_patch: MagicMock,
